+++ src/createComponent.tsx
@@ -1,17 +1,15 @@
 import app from './app';
 
 const cache = {}
-
-export default function (componentClass, state) {
-  let id = state && state['id'];
-  let key = state && (state['key'] || state['id']);
+app.on('get-component', o => o.components = cache);
+export default function (componentClass, props) {
+  let id = props && props['id'];
   if (!id) id = `_${componentClass.name}_${Date.now()}`;
-
-  const component = key && cache[key] ? cache[key] :
-  (cache[key] = new componentClass(state).mount(id));
-  if (component.mounted) setTimeout(()=>component.mounted(state), 0);
-
-  return <div id={id} _component={component}>
+  const component = (id && cache[id]) ? cache[id] :
+  (cache[id] = new componentClass(props).mount(id));
+  if (component.mounted) setTimeout(() => component.mounted(props), 0);
+  if (component.rendered) setTimeout(() => component.rendered(component.state), 0);
+  return <div id={id}>
     {component.view && component.view(component.state)}
   </div>
 }